symphonia_utils_xiph::flac::metadata
Struct StreamInfo

pub struct StreamInfo {
    pub block_len_min: u16,
    pub block_len_max: u16,
    pub frame_byte_len_min: u32,
    pub frame_byte_len_max: u32,
    pub sample_rate: u32,
    pub channels: Channels,
    pub bits_per_sample: u32,
    pub n_samples: Option<u64>,
    pub md5: Option<[u8; 16]>,
}

Fields
block_len_min: u16

The minimum and maximum number of decoded samples per block of audio.

block_len_max: u16frame_byte_len_min: u32

The minimum and maximum byte length of an encoded block (frame) of audio. Either value may be 0 if unknown.

frame_byte_len_max: u32sample_rate: u32

The sample rate in Hz.

channels: Channels

The channel mask.

bits_per_sample: u32

The number of bits per sample of the stream.

n_samples: Option<u64>

The total number of samples in the stream, if available.

md5: Option<[u8; 16]>

The MD5 hash value of the decoded audio.

Implementations

impl StreamInfo

pub fn read<B: ReadBytes>(reader: &mut B) -> Result<StreamInfo>

Read a stream information block.

pub fn is_valid_size(size: u64) -> bool

Check if the size is valid for a stream information block.